export di for compatibility with amd modules
