+++ Types/_chain/Abstract.ts
@@ -12,7 +12,7 @@
 import {DestroyableMixin} from '../entity';
 import {IEnumerable} from '../collection';
 import {object} from '../util';
-import di from '../di';
+import di from '../_di';
 import IEnumerator from '../_collection/IEnumerator';
 import Zipped from './Zipped';
 import Mapped from './Mapped';
