+++ Types/_display/FlagsItem.ts
@@ -8,7 +8,7 @@
  */
 
 import CollectionItem from './CollectionItem';
-import di from '../di';
+import di from '../_di';
 
 export default class FlagsItem extends CollectionItem /** @lends Types/Display/FlagsItem.prototype */{
    isSelected(): boolean {
